+++ tests/server/util/util.js
@@ -695,6 +695,17 @@ describe('util', function() {
         assert.equal(validatedObject.dateProperty, date.getTime());
       });
 
+      it('should be able to validate a timestamp as a string', function() {
+        var date = new Date();
+        var validatedObject = util.shallowValidateObject({
+          dateProperty: String(date.getTime())
+        }, {
+          dateProperty: {type: 'date'}
+        });
+
+        assert.strictEqual(validatedObject.dateProperty, date.getTime());
+      });
+
       it('should be able to validate a date', function() {
         var date = new Date('03/22/2016');
         var validatedObject = util.shallowValidateObject({